test: 리팩토링 테스트 반영 및 테스트 환경 설정
